// src/generated/models/Afb_journalauditsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Afb_journalauditsafb_typedaction = {
  0: 'Cr_ation',
  1: 'Modification',
  2: 'Export',
  3: 'Connexion',
  747010001: 'Lecture',
  747010002: 'Rejet',
  747010003: 'Suppression',
  747010004: 'Validation'
} as const;
export type Afb_journalauditsafb_typedaction = keyof typeof Afb_journalauditsafb_typedaction;
export const Afb_journalauditsstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Afb_journalauditsstatecode = keyof typeof Afb_journalauditsstatecode;
export const Afb_journalauditsstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Afb_journalauditsstatuscode = keyof typeof Afb_journalauditsstatuscode;

export interface Afb_journalauditsBase {
  afb_adresseipsource?: string;
  afb_agentutilisateur?: string;
  afb_anciennevaleur?: string;
  "afb_auteurdelamodification@odata.bind"?: string;
  afb_entitemodifiee: string;
  afb_guiddelenregistrement: string;
  afb_horodatage: string;
  afb_identifiantdujournal: string;
  afb_journalauditid: string;
  afb_nouvellevaleur?: string;
  afb_typedaction: Afb_journalauditsafb_typedaction;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Afb_journalauditsstatecode;
  statuscode?: Afb_journalauditsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Afb_journalaudits extends Afb_journalauditsBase {
  afb_auteurdelamodificationname?: string;
  afb_typedactionname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  afb_auteurdelamodification?: object;
  _afb_auteurdelamodification_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}